custom game creation form to handle api call
